Ecommerce/home/serializer.py
from rest_framework import serializers
from home.models import *


class ProductSerializer(serializers.ModelSerializer):
    class Meta:
        model = Slider
        fields = "__all__"


class CategorySerializer(serializers.ModelSerializer):
    class Meta:
        model = Category
        fields = "__all__"
